requirements.txt
alembic==0.9.6
bcrypt
click==6.7
coverage==4.4.2
eventlet==0.31.0
flask>=0.12.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.1.1
flask-restx==0.2.0
pip==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.7.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2>=2.10.1
jsonschema==2.6.0
Mako==1.1.2
psycopg2-binary
pycparser==2.18
PyJWT==1.7.1
python-dateutil==2.6.1
pytz==2017.3
selenium==3.8.1
six==1.11.0
SQLAlchemy>=1.3.0
Werkzeug==0.15.5
face_recognition
opencv-python
numpy
imutils